Finished Projects
Made it so that all interests will be displayed on the interest array.

Made it so that all the days are displayed in English and French in the DaysArray.php file

Finally, I made it so the IsEven.php will generate random numbers then decide whether they are even or not.