PHP Updates
Made a file that will load in PHP headers and footers.

Made a file to show conditonal scripts.

Made a dice roll game.

Made a tutorial to help teacher Jaggar how to do the last project.

Made a file that will decide whether randomly generated numbers are odd or not.

Made a file to document variable scope.

Made a file that will convert fahrenheit to celsius.

Made a file to show off while loop logic.